WIP - Support truly async Scala.js test
Previously, we didn't really support properly async Scala.js tests. Now,
the evaluation of test cases has been refactored so that the sbt test
runner has control over evaluation of the test suite in a `Future[Unit]`.

This refactoring required some public APIs but it also cleared up some
parts that I was not too happy about. The implementations of the `.fail`
and `.flaky` modifiers are much cleaner now.