Experiment with Dotty extension methods for TestOptions.
This commit is a proof-of-concept how we could avoid implicit
conversions on `TestOptions` for usage like `test("name".only)`.
We still need to enable implicit conversions to compile the MUnit codebase
since it uses a lot of helper methods like
`def check(options: TestOptions)` where we don't add a method overload.